backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.44
botocore==1.42.44
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Deprecated==1.3.1
distro==1.9.0
fastapi==0.128.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jmespath==1.1.0
limits==5.8.0
openai==2.17.0
orjson==3.11.7
packaging==24.2
pinecone==8.0.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wrapt==2.1.1
mangum
langchain-text-splitters
langchain-core==0.3.83
langchain-text-splitters==0.3.11psycopg2-binary
